Rename test utility namespace
